Refactored the old mai loop to be on core1 in case running our processing loop being on core1 was causing memory issues and prints not working right. This was me misunderstanding what was going on. It doesn't matter.
Using another approach to command 0x71 (and 0x11) using a rom array I found in another source base.

Suspect an electrical issue at this point. Depending on how I wiggle it while it's plugged in I may or may not get stuck on a black screen, the sega dreamcast logo, or the dashboard stuck on checking disc.